done permissions

+++ packages/plugin-filemanager-api/src/graphql/resolvers/mutations.ts
@@ -1,4 +1,4 @@
-import { checkPermission } from '@erxes/api-utils/src/permissions';
+// import { checkPermission } from '@erxes/api-utils/src/permissions';
 import { IContext } from '../../connectionResolver';
 
 const mutations = {
@@ -11,7 +11,13 @@ const mutations = {
     });
   },
 
-  filemanagerFolderRemove(_root, { _id }, { models }: IContext) {
+  async filemanagerFolderRemove(_root, { _id }, { models, user }: IContext) {
+    const folder = await models.Folders.getFolder({ _id });
+
+    if (folder.createdUserId !== user._id) {
+      throw new Error('Access denied');
+    }
+
     return models.Folders.deleteOne({ _id });
   },
 
@@ -19,12 +25,41 @@ const mutations = {
     return models.Files.saveFile({ doc });
   },
 
-  filemanagerFileRemove(_root, { _id }, { models }: IContext) {
+  async filemanagerFileRemove(_root, { _id }, { models, user }: IContext) {
+    const file = await models.Files.getFile({ _id });
+
+    if (file.createdUserId !== user._id) {
+      throw new Error('Access denied');
+    }
+
     return models.Files.deleteOne({ _id });
+  },
+
+  async filemanagerChangePermission(
+    _root,
+    { type, _id, userIds, unitId },
+    { models, user }: IContext
+  ) {
+    let collection: any = models.Folders;
+
+    if (type === 'file') {
+      collection = models.Files;
+    }
+
+    const object = await collection.findOne({ _id });
+
+    if (object.createdUserId !== user._id) {
+      throw new Error('Permission denied');
+    }
+
+    return collection.update(
+      { _id },
+      { $set: { permissionConfig: { userIds, unitId } } }
+    );
   }
 };
 
-checkPermission(mutations, 'filemanagerFolderSave', 'manageFilemanager');
-checkPermission(mutations, 'filemanagerFolderRemove', 'manageFilemanager');
+// checkPermission(mutations, 'filemanagerFolderSave', 'manageFilemanager');
+// checkPermission(mutations, 'filemanagerFolderRemove', 'manageFilemanager');
 
 export default mutations;
